+++ spec/lib/js_regex/converter/property_converter_spec.rb
@@ -10,7 +10,8 @@ describe JsRegex::Converter::PropertyConverter do
 
   it 'translates the negated \p{^...} property style' do
     expect(/\p{^ascii}/).to\
-    become(/[^\x00-\x7F]/).and keep_matching('añB', with_results: %w[ñ])
+    become('(?:[\x80-\uD7FF\uE000-\uFFFF]|[\uD800-\uDBFF][\uDC00-\uDFFF])')
+      .and keep_matching('añB😁', with_results: %w[ñ 😁])
   end
 
   it 'translates the double-negated \P{^...} property style' do
@@ -18,19 +19,9 @@ describe JsRegex::Converter::PropertyConverter do
     become(/[\x00-\x7F]/).and keep_matching('añB', with_results: %w[a B])
   end
 
-  it 'drops astral plane properties negated with \p{^ with warning' do
-    expect(/\p{^Deseret}/).to\
-    become(//).with_warning('astral plane negation by property')
-  end
-
-  it 'drops astral plane properties negated with \P with warning' do
-    expect(/\P{Deseret}/).to\
-    become(//).with_warning('astral plane negation by property')
-  end
-
   it 'translates abbreviated properties' do
     expect(/\p{cc}/).to\
-    become(double(source: '[\x00-\x1F\x7F-\x9F]'))
+    become('[\x00-\x1F\x7F-\x9F]')
       .and keep_matching('A B', with_results: %w[ ])
   end
 
@@ -42,34 +33,11 @@ describe JsRegex::Converter::PropertyConverter do
   it 'does not use case-insensitive substitutions if everything is i anyway' do
     result = JsRegex.new(/1\p{lower}2/i)
     expect(result.source).not_to include 'A-Z'
+    expect(result.warnings).to be_empty
   end
 
-  it 'cuts of large astral plane shares of properties with warning' do
-    expect(JsRegex::Converter).to(receive(:in_surrogate_pair_limit?) do |&block|
-      expect(block.call).to eq 1048576
-    end).and_return(false)
-    expect(/\p{Any}/).to\
-    become(/[\x00-\uFFFF]/).with_warning('large astral plane match of property')
-  end
-
-  it 'drops too large astral plane properties with warning',
-    if: (Gem::Version.new(RUBY_VERSION.dup) >= Gem::Version.new('2.4.1')) do
-    # this should concern little more than the few astral plane scripts
-    # supported by Ruby, but it is also a good precaution if spacy new
-    # properties are added in the future.
-    expect(JsRegex::Converter)
-      .to receive(:in_surrogate_pair_limit?)
-      .and_call_original
-    expect(Regexp.new('\p{SignWriting}'))
-      .to become(//)
-      .with_warning('large astral plane match of property')
-  end
-
-  it 'allows large astral plane properties if the limit allows it',
-    if: (Gem::Version.new(RUBY_VERSION.dup) >= Gem::Version.new('2.4.1')) do
-    expect(JsRegex::Converter)
-      .to receive(:in_surrogate_pair_limit?)
-      .and_return(true)
-    expect(Regexp.new('a\p{SignWriting}b')).to keep_matching('a𝠔b')
+  it 'warns for nested case-sensitive properties' do
+    expect(/(?-i:\p{upper})/i)
+      .to generate_warning('nested case-sensitive property')
   end
 end