+++ spec/lib/js_regex/converter/type_converter_spec.rb
@@ -32,12 +32,12 @@ describe JsRegex::Converter::TypeConverter do
 
   it 'substitutes the hex type "\h" with an equivalent set' do
     expect(/\h+/).to\
-    become(/[0-9A-Fa-f]+/).and(keep_matching('FF__FF', with_results: %w[FF FF]))
+    become(/[0-9A-Fa-f]+/).and keep_matching('f').and keep_not_matching('x')
   end
 
   it 'substitutes the nonhex type "\H" with an equivalent set' do
     expect(/\H+/).to\
-    become(/[^0-9A-Fa-f]+/).and keep_matching('FFxy66z', with_results: %w[xy z])
+    become(/[^0-9A-Fa-f]+/).and keep_matching('x').and keep_not_matching('f')
   end
 
   it 'substitutes the generic linebreak type "\R"' do